+++ openinfer/src/ops/cpu/mul/mod.rs
@@ -5,7 +5,7 @@ use crate::timer::Timer;
 
 pub mod registry;
 
-pub fn mul_i8(a: &[i8], b: &[i8], thread_id: u32) -> Result<Vec<i8>> {
+pub fn mul_i8(a: &[i8], b: &[i8], thread_id: usize) -> Result<Vec<i8>> {
     if a.len() != b.len() {
         return Err(anyhow!("mul op shape mismatch"));
     }
@@ -18,7 +18,7 @@ pub fn mul_i8(a: &[i8], b: &[i8], thread_id: u32) -> Result<Vec<i8>> {
     Ok(out)
 }
 
-pub fn mul_i16(a: &[i16], b: &[i16], thread_id: u32) -> Result<Vec<i16>> {
+pub fn mul_i16(a: &[i16], b: &[i16], thread_id: usize) -> Result<Vec<i16>> {
     if a.len() != b.len() {
         return Err(anyhow!("mul op shape mismatch"));
     }
@@ -31,7 +31,7 @@ pub fn mul_i16(a: &[i16], b: &[i16], thread_id: u32) -> Result<Vec<i16>> {
     Ok(out)
 }
 
-pub fn mul_f32(a: &[f32], b: &[f32], thread_id: u32) -> Result<Vec<f32>> {
+pub fn mul_f32(a: &[f32], b: &[f32], thread_id: usize) -> Result<Vec<f32>> {
     if a.len() != b.len() {
         return Err(anyhow!("mul op shape mismatch"));
     }
@@ -44,7 +44,7 @@ pub fn mul_f32(a: &[f32], b: &[f32], thread_id: u32) -> Result<Vec<f32>> {
     Ok(out)
 }
 
-pub fn mul_f64(a: &[f64], b: &[f64], thread_id: u32) -> Result<Vec<f64>> {
+pub fn mul_f64(a: &[f64], b: &[f64], thread_id: usize) -> Result<Vec<f64>> {
     if a.len() != b.len() {
         return Err(anyhow!("mul op shape mismatch"));
     }
@@ -57,7 +57,7 @@ pub fn mul_f64(a: &[f64], b: &[f64], thread_id: u32) -> Result<Vec<f64>> {
     Ok(out)
 }
 
-pub fn mul_u8(a: &[u8], b: &[u8], thread_id: u32) -> Result<Vec<u8>> {
+pub fn mul_u8(a: &[u8], b: &[u8], thread_id: usize) -> Result<Vec<u8>> {
     if a.len() != b.len() {
         return Err(anyhow!("mul op shape mismatch"));
     }
@@ -70,7 +70,7 @@ pub fn mul_u8(a: &[u8], b: &[u8], thread_id: u32) -> Result<Vec<u8>> {
     Ok(out)
 }
 
-pub fn mul_u16(a: &[u16], b: &[u16], thread_id: u32) -> Result<Vec<u16>> {
+pub fn mul_u16(a: &[u16], b: &[u16], thread_id: usize) -> Result<Vec<u16>> {
     if a.len() != b.len() {
         return Err(anyhow!("mul op shape mismatch"));
     }
@@ -83,7 +83,7 @@ pub fn mul_u16(a: &[u16], b: &[u16], thread_id: u32) -> Result<Vec<u16>> {
     Ok(out)
 }
 
-pub fn mul_i32(a: &[i32], b: &[i32], thread_id: u32) -> Result<Vec<i32>> {
+pub fn mul_i32(a: &[i32], b: &[i32], thread_id: usize) -> Result<Vec<i32>> {
     if a.len() != b.len() {
         return Err(anyhow!("mul op shape mismatch"));
     }
@@ -96,7 +96,7 @@ pub fn mul_i32(a: &[i32], b: &[i32], thread_id: u32) -> Result<Vec<i32>> {
     Ok(out)
 }
 
-pub fn mul_i64(a: &[i64], b: &[i64], thread_id: u32) -> Result<Vec<i64>> {
+pub fn mul_i64(a: &[i64], b: &[i64], thread_id: usize) -> Result<Vec<i64>> {
     if a.len() != b.len() {
         return Err(anyhow!("mul op shape mismatch"));
     }
@@ -109,7 +109,7 @@ pub fn mul_i64(a: &[i64], b: &[i64], thread_id: u32) -> Result<Vec<i64>> {
     Ok(out)
 }
 
-pub fn mul_u32(a: &[u32], b: &[u32], thread_id: u32) -> Result<Vec<u32>> {
+pub fn mul_u32(a: &[u32], b: &[u32], thread_id: usize) -> Result<Vec<u32>> {
     if a.len() != b.len() {
         return Err(anyhow!("mul op shape mismatch"));
     }
@@ -122,7 +122,7 @@ pub fn mul_u32(a: &[u32], b: &[u32], thread_id: u32) -> Result<Vec<u32>> {
     Ok(out)
 }
 
-pub fn mul_u64(a: &[u64], b: &[u64], thread_id: u32) -> Result<Vec<u64>> {
+pub fn mul_u64(a: &[u64], b: &[u64], thread_id: usize) -> Result<Vec<u64>> {
     if a.len() != b.len() {
         return Err(anyhow!("mul op shape mismatch"));
     }
@@ -135,7 +135,7 @@ pub fn mul_u64(a: &[u64], b: &[u64], thread_id: u32) -> Result<Vec<u64>> {
     Ok(out)
 }
 
-pub fn mul_bool(a: &[bool], b: &[bool], thread_id: u32) -> Result<Vec<bool>> {
+pub fn mul_bool(a: &[bool], b: &[bool], thread_id: usize) -> Result<Vec<bool>> {
     if a.len() != b.len() {
         return Err(anyhow!("mul op shape mismatch"));
     }
@@ -148,14 +148,14 @@ pub fn mul_bool(a: &[bool], b: &[bool], thread_id: u32) -> Result<Vec<bool>> {
     Ok(out)
 }
 
-pub fn mul_bitset(_a: &[Bitset], _b: &[Bitset], thread_id: u32) -> Result<Vec<Bitset>> {
+pub fn mul_bitset(_a: &[Bitset], _b: &[Bitset], thread_id: usize) -> Result<Vec<Bitset>> {
     println!("cpu mul_bitset: stub");
     Timer::start(thread_id);
     Timer::stop(thread_id);
     Ok(Vec::new())
 }
 
-pub fn mul_f16(_a: &[F16], _b: &[F16], thread_id: u32) -> Result<Vec<F16>> {
+pub fn mul_f16(_a: &[F16], _b: &[F16], thread_id: usize) -> Result<Vec<F16>> {
     println!("cpu mul_f16: stub");
     Timer::start(thread_id);
     Timer::stop(thread_id);
