+++ openinfer/src/ops/cpu_avx/add/mod.rs
@@ -10,7 +10,7 @@ use crate::timer::Timer;
 
 pub mod registry;
 
-pub fn add_f32(a: &[f32], b: &[f32], thread_id: u32) -> Result<Vec<f32>> {
+pub fn add_f32(a: &[f32], b: &[f32], thread_id: usize) -> Result<Vec<f32>> {
     if a.len() != b.len() {
         return Err(anyhow!("add op shape mismatch"));
     }
@@ -36,7 +36,7 @@ pub fn add_f32(a: &[f32], b: &[f32], thread_id: u32) -> Result<Vec<f32>> {
     Ok(out)
 }
 
-pub fn add_i8(a: &[i8], b: &[i8], thread_id: u32) -> Result<Vec<i8>> {
+pub fn add_i8(a: &[i8], b: &[i8], thread_id: usize) -> Result<Vec<i8>> {
     if a.len() != b.len() {
         return Err(anyhow!("add op shape mismatch"));
     }
@@ -62,7 +62,7 @@ pub fn add_i8(a: &[i8], b: &[i8], thread_id: u32) -> Result<Vec<i8>> {
     Ok(out)
 }
 
-pub fn add_i16(a: &[i16], b: &[i16], thread_id: u32) -> Result<Vec<i16>> {
+pub fn add_i16(a: &[i16], b: &[i16], thread_id: usize) -> Result<Vec<i16>> {
     if a.len() != b.len() {
         return Err(anyhow!("add op shape mismatch"));
     }
@@ -88,7 +88,7 @@ pub fn add_i16(a: &[i16], b: &[i16], thread_id: u32) -> Result<Vec<i16>> {
     Ok(out)
 }
 
-pub fn add_f64(a: &[f64], b: &[f64], thread_id: u32) -> Result<Vec<f64>> {
+pub fn add_f64(a: &[f64], b: &[f64], thread_id: usize) -> Result<Vec<f64>> {
     if a.len() != b.len() {
         return Err(anyhow!("add op shape mismatch"));
     }
@@ -114,7 +114,7 @@ pub fn add_f64(a: &[f64], b: &[f64], thread_id: u32) -> Result<Vec<f64>> {
     Ok(out)
 }
 
-pub fn add_u8(a: &[u8], b: &[u8], thread_id: u32) -> Result<Vec<u8>> {
+pub fn add_u8(a: &[u8], b: &[u8], thread_id: usize) -> Result<Vec<u8>> {
     if a.len() != b.len() {
         return Err(anyhow!("add op shape mismatch"));
     }
@@ -140,7 +140,7 @@ pub fn add_u8(a: &[u8], b: &[u8], thread_id: u32) -> Result<Vec<u8>> {
     Ok(out)
 }
 
-pub fn add_u16(a: &[u16], b: &[u16], thread_id: u32) -> Result<Vec<u16>> {
+pub fn add_u16(a: &[u16], b: &[u16], thread_id: usize) -> Result<Vec<u16>> {
     if a.len() != b.len() {
         return Err(anyhow!("add op shape mismatch"));
     }
@@ -166,7 +166,7 @@ pub fn add_u16(a: &[u16], b: &[u16], thread_id: u32) -> Result<Vec<u16>> {
     Ok(out)
 }
 
-pub fn add_i32(a: &[i32], b: &[i32], thread_id: u32) -> Result<Vec<i32>> {
+pub fn add_i32(a: &[i32], b: &[i32], thread_id: usize) -> Result<Vec<i32>> {
     if a.len() != b.len() {
         return Err(anyhow!("add op shape mismatch"));
     }
@@ -192,7 +192,7 @@ pub fn add_i32(a: &[i32], b: &[i32], thread_id: u32) -> Result<Vec<i32>> {
     Ok(out)
 }
 
-pub fn add_i64(a: &[i64], b: &[i64], thread_id: u32) -> Result<Vec<i64>> {
+pub fn add_i64(a: &[i64], b: &[i64], thread_id: usize) -> Result<Vec<i64>> {
     if a.len() != b.len() {
         return Err(anyhow!("add op shape mismatch"));
     }
@@ -218,7 +218,7 @@ pub fn add_i64(a: &[i64], b: &[i64], thread_id: u32) -> Result<Vec<i64>> {
     Ok(out)
 }
 
-pub fn add_u32(a: &[u32], b: &[u32], thread_id: u32) -> Result<Vec<u32>> {
+pub fn add_u32(a: &[u32], b: &[u32], thread_id: usize) -> Result<Vec<u32>> {
     if a.len() != b.len() {
         return Err(anyhow!("add op shape mismatch"));
     }
@@ -244,7 +244,7 @@ pub fn add_u32(a: &[u32], b: &[u32], thread_id: u32) -> Result<Vec<u32>> {
     Ok(out)
 }
 
-pub fn add_u64(a: &[u64], b: &[u64], thread_id: u32) -> Result<Vec<u64>> {
+pub fn add_u64(a: &[u64], b: &[u64], thread_id: usize) -> Result<Vec<u64>> {
     if a.len() != b.len() {
         return Err(anyhow!("add op shape mismatch"));
     }
@@ -270,7 +270,7 @@ pub fn add_u64(a: &[u64], b: &[u64], thread_id: u32) -> Result<Vec<u64>> {
     Ok(out)
 }
 
-pub fn add_bool(a: &[bool], b: &[bool], thread_id: u32) -> Result<Vec<bool>> {
+pub fn add_bool(a: &[bool], b: &[bool], thread_id: usize) -> Result<Vec<bool>> {
     if a.len() != b.len() {
         return Err(anyhow!("add op shape mismatch"));
     }
@@ -296,10 +296,10 @@ pub fn add_bool(a: &[bool], b: &[bool], thread_id: u32) -> Result<Vec<bool>> {
     Ok(out)
 }
 
-pub fn add_bitset(a: &[Bitset], b: &[Bitset], thread_id: u32) -> Result<Vec<Bitset>> {
+pub fn add_bitset(a: &[Bitset], b: &[Bitset], thread_id: usize) -> Result<Vec<Bitset>> {
     crate::ops::cpu::add::add_bitset(a, b, thread_id)
 }
 
-pub fn add_f16(a: &[F16], b: &[F16], thread_id: u32) -> Result<Vec<F16>> {
+pub fn add_f16(a: &[F16], b: &[F16], thread_id: usize) -> Result<Vec<F16>> {
     crate::ops::cpu::add::add_f16(a, b, thread_id)
 }
